[td] Reload data when saving action
